+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,15 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [3.13.0-beta.2] - 2025/11/2
+
+- **Added**
+  - Better AvatarPedestal に以下を追加しました。
+    - Dynamics Parameter Transfer
+      - Animator に付ける事で、子孫オブジェクトにある全ての Physbone と Contact Receiver Infomation を走査し、その状態を Animator にパラメーターとして渡せるコンポーネントです。
+    - Contact Receiver Infomation
+      - Dynamics Parameter Transfer から Contact Receiver の状態を読み取るための補助コンポーネントです。Contact Receiver にセットで付けます。
+
 ## [3.13.0-beta.1] - 2025/10/28
 
 - **Added**
@@ -749,6 +758,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[3.13.0-beta.2]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.13.0-beta.2
 [3.13.0-beta.1]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.13.0-beta.1
 [3.12.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.12.0
 [3.11.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.11.0